requirements.txt
pandas
openai
python-dotenv
llama-cpp-python
langchain_community
langchain_core
langserve[all]